working.

+++ src/core/pathid_set.h
@@ -14,13 +14,13 @@ typedef struct QUIC_PATHID_TYPE_INFO {
     // The largest MAX_{CLIENT, SERVER}_PATHS value indicated to the peer. This MUST not ever
     // decrease once the connection has started.
     //
-    uint64_t MaxTotalPathIDCount;
+    uint32_t MaxTotalPathIDCount;
 
     //
     // The total number of path ids that have been opened. Includes any path ids
     // that have been closed as well.
     //
-    uint64_t TotalPathIDCount;
+    uint32_t TotalPathIDCount;
 
     //
     // The maximum number of simultaneous open path ids allowed.
@@ -65,7 +65,7 @@ typedef struct QUIC_PATHID_SET {
             CXPLAT_HASHTABLE* Table;
         } HASH;
     };
-    
+
 } QUIC_PATHID_SET;
 
 //
